Fix local testing with test_extensions.rb, update js combine and transpile [skip ci]
